fixed component Post

+++ src/components/Home.js
@@ -2,6 +2,7 @@ import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
 import axios from 'axios'
 import { connect } from 'react-redux';
+import swal from 'sweetalert';
 
 class Home extends Component {
 
@@ -11,7 +12,7 @@ class Home extends Component {
           userId:'',  
           title: '',  
           _id: '',
-          posts: [ ],
+          onePost: [ ],
         };
     }
     
@@ -21,14 +22,33 @@ class Home extends Component {
 
     getPosts = async () =>{
         const res = await axios.get('https://jsonplaceholder.typicode.com/posts');
-        this.setState({posts: res.data});
+        this.props.setPosts(res.data)
     }
 
     deletePost = async (postId) =>{
-        await axios.delete('https://jsonplaceholder.typicode.com/posts/' + postId);
-        this.props.dispatch({type:'DELETE_POST', id: postId})
-    }
+        swal({
+            title: 'Delete Post',
+            text: 'Are you sure you want to delete this post?',
+            icon: 'error',
+            buttons: ['No', 'Yes']
+        }).then(async respuesta =>{
+            if (respuesta) {
+            await axios.delete('https://jsonplaceholder.typicode.com/posts/' + postId);
+            this.props.deletePost(postId)
+            swal({text: 'The post was successfully deleted',
+            icon: 'success'})
+        }
+    })
+  }
 
+    detailsPost = async (postId) =>{
+        const dataPost = await axios.get('https://jsonplaceholder.typicode.com/posts/' + postId);
+        this.setState({onePost: dataPost.data});
+        swal(
+            {title:`${this.state.onePost.title}`,
+            text: `${this.state.onePost.body}`,
+            icon: 'info'})
+    }
     render(){
         return(
             <div className="container">
@@ -44,16 +64,15 @@ class Home extends Component {
                             </thead>
                             <tbody>
                             { 
-                            this.state.posts.map(post => {
+                            this.props.posts.map(post => {
                             return (
-                                <tr>
+                                <tr key={post.id}>
                                     <td>{post.userId}</td>
                                     <td>{post.title}</td>
                                     <td>
-                                    <Link to={"/details/" + post.id} className="btn btn-primary" style={{margin: '4px'}}>
-                                        <i className="material-icons"> 
-                                            info</i>
-                                    </Link>      
+                                    <button onClick={() => this.detailsPost(post.id)} className="btn btn-primary" style={{margin: '4px'}}>
+                                        details
+                                    </button>      
                                     <Link to={"/edit/" + post.id} className="btn btn-warning" style={{margin: '4px'}}>
                                         <i className="material-icons">
                                             edit</i>
@@ -73,5 +92,19 @@ class Home extends Component {
             </div>  
         )
     }
-}       
-export default connect()(Home);
+}            
+
+const mapStateToProps = (state) => {
+    return {
+        posts: state
+    }
+}
+
+const mapDispatchToProps = dispatch => {
+    return {
+        setPosts: posts => dispatch({ type: 'GET_POSTS', payload: posts }),
+        deletePost: postId => dispatch({ type:'DELETE_POST', payload: postId }),
+    }
+}
+
+export default connect(mapStateToProps, mapDispatchToProps)(Home);
